Don't constrain epochlength to be an int, allow floats

+++ pymento_meg/cli.py
@@ -382,7 +382,7 @@ def epoch_and_clean():
                            bidsdir=args.bidsdir,
                            datadir=args.datadir,
                            derivdir=args.derivdir,
-                           epochlength=int(args.epochlength),
+                           epochlength=float(args.epochlength),
                            eventid=eventid)
